fixed onto file path stuff and remove ann
