+++ data/owl/annett-o-schema.owl
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:ddf795fa9f999f5f3c75a5954fef8d94b8a9d14e0ae69cda629ae3e4fc8b1fb3
+size 100868
